Update 1275 Find Winner on a Tic Tac Toe Game (#76)
Tic Tac Toe solution used `break` inside a loop when the first cell of the column or row was empty. I believe this will result in it skipping all other rows/columns when it reaches a null cell. Instead of using `break` I think it should be `continue` instead.

Tested locally and `continue` works for 3 vertical in the last column, where `break` fails to find the winner.